requirements.txt
Flask
flask-cors
pandas
python-docx
openpyxl
Werkzeug
gunicorn
